Allow usage of shinyTree in modules
Ids in shiny modules are namespaced by prepending the modules id and a
dash "-". The shinyTree id is used "as is" as a JavaScript variable
name in the head of the HTML file, which causes a bug, because the dash
is an invalid character for variable names.

The dash was consequently replaced by an underscore "_" in the HTML
generating R function as well as in the JavaScript file which
references this variable.