Add default text value when creating a new node
By default the text value is an empty string, which causes the app to
crash when using the `tree` parser.

I add the placeholder as the default value.

This default value will be erased once the user confirms the name of the
new node, which will always happen, voluntarily or not (the name gets
confirmed as soon as the focus is not on the text field anymore, e.g. if
the user clicks somewhere else).